requirements.txt
openai
python-dotenv
uvicorn
fastapi
llama-index-core
pymongo
python-multipart
gunicorn
pypdf
boto3